fix(sonar): real dshot value-truncation bug + at-source smell remediation
Remediate SonarCloud findings at source (the design-conflict style rules are
handled separately by tools/sonar/accept_style_conflicts.py):

- fix(dshot): DSHOT_THROTTLE_MAX / DSHOT_VALUE_MAX were `uint8_t` = 2047, which
  truncates to 255 (S5276, value-changing narrowing). The DShot value field is
  11-bit (0..2047), so these are now `uint16_t`. Latent (no current caller) but a
  wrong public constant a `value <= DSHOT_VALUE_MAX` check would have trusted.
- S1659: split a two-in-one declaration in quic_frame.cpp (one decl per line).
- S1110: drop redundant parentheses in mlkem.cpp.
- S5416: modernize the callback/struct typedefs to `using` / plain `struct`
  across relay, dnc, smb, forward, wal, dbm, docstore, sqlite, ads (17 sites).
  Purely syntactic - the raw-function-pointer design (S5205) is unchanged.
- tools/sonar: add S5028 (macros-define-constants) to the accept list. The policy
  doc already documents it as by-design (the whole #if-gated config system is
  `#define`), but the script omitted it, so 66 issues never got marked.

Verified: native_dshot, native_quic_frame, native_pqc, native_relay, native_dnc,
native_smb, native_forward, native_wal, native_dbm, native_docstore,
native_sqlite, native_ads all green; accept script still compiles.

+++ src/services/dshot/dshot.h
@@ -43,9 +43,11 @@ struct DshotCmd
     static constexpr uint8_t DSHOT_CMD_3D_MODE_ON = 10;      ///< enable bidirectional 3D mode (send 6x).
     static constexpr uint8_t DSHOT_CMD_SETTINGS_REQUEST = 11;
     static constexpr uint8_t DSHOT_CMD_SAVE_SETTINGS = 12; ///< persist settings (send 6x).
-    static constexpr uint8_t DSHOT_THROTTLE_MIN = 48;      ///< first real throttle step.
-    static constexpr uint8_t DSHOT_THROTTLE_MAX = 2047;    ///< last throttle step (2000 steps of resolution).
-    static constexpr uint8_t DSHOT_VALUE_MAX = 2047;       ///< widest value the 11-bit field holds.
+    // The throttle / value domain is the 11-bit value field (0..2047), so these need a 16-bit type -
+    // uint8_t would silently truncate 2047 to 255.
+    static constexpr uint16_t DSHOT_THROTTLE_MIN = 48;   ///< first real throttle step.
+    static constexpr uint16_t DSHOT_THROTTLE_MAX = 2047; ///< last throttle step (2000 steps of resolution).
+    static constexpr uint16_t DSHOT_VALUE_MAX = 2047;    ///< widest value the 11-bit field holds.
 };
 
 /**
